Sun, Sep 4, 2022, 10:18 PM +00:00

+++ week1/calculator.c
@@ -8,9 +8,9 @@ int main(void)
     float y = get_float("y: ");
 
     //Divide x by y
-    float z = x / y;
+    float z = x / (float) y;
 
     //perform
-    printf(" = %.2f\n", z);
+    printf(" = %.50f\n", z);
 
 }
